+++ server/src/util/types/index.ts
@@ -1,27 +1,12 @@
-//  consideration: as schema library, zod could validate types with built-in codes.
+/*
+  [DISCLAIMER]
 
-export type type_dept_row = {
-  dept_name: string;
-  dept_capacity: number;
-  importance_weight: number;
-  is_active: boolean;
-};
+  Types is the centralised storage of type managements for different section.
+  These types specified those frequent used and complex structured types for
+  better maintenance.
+*/
 
-export type type_staff_row = {
-  first_name: string;
-  last_name: string;
-  gender: enum_gender;
-  work_position: string;
-  work_grade: enum_staff_grade;
-  work_email: string | null;
-  work_ext: string | null;
-  dept_id: number | null;
-  date_hired: Date | null;
-  date_quit: Date | null;
-  is_active: boolean;
-};
-// ＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋＋÷
-
+//  tags: departments, staff
 export type TSchemaBase = {
   _id: number;
   created_at: Date;